chore: Initial pass at cleanup of initializeAgent
