+++ ui/desktop/openapi.json
@@ -10,7 +10,7 @@
     "license": {
       "name": "Apache-2.0"
     },
-    "version": "1.0.15"
+    "version": "1.0.16"
   },
   "paths": {
     "/config": {
